Reverted 7278376ab77651e540e39552384ad9677e32ff7e.
Since the removal of rails/init.rb, Rails fails to load the helpers.

After investigation, it seems that Rails 2.3.5 handles rails/init.rb in a special way
creating an instance of class GemPlugin < Plugin.
Rails 3 doesn't support that feature, I would have to create a Railtie.

So, let keep this branch compatible with Rails 2.3.x and create a new branch for Rails 3.